+++ Drawing/DrawOps/ReplaceDrawOp.cs
@@ -30,7 +30,7 @@ namespace MCGalaxy.Drawing.Ops {
         
         public override string Name { get { return "Replace"; } }
         
-        public override int GetBlocksAffected(Level lvl, Vec3U16[] marks) {
+        public override long GetBlocksAffected(Level lvl, Vec3U16[] marks) {
             Vec3U16 p1 = marks[0], p2 = marks[1];
             return (p2.X - p1.X + 1) * (p2.Y - p1.Y + 1) * (p2.Z - p1.Z + 1);
         }
@@ -60,7 +60,7 @@ namespace MCGalaxy.Drawing.Ops {
         
         public override string Name { get { return "ReplaceNot"; } }
         
-        public override int GetBlocksAffected(Level lvl, Vec3U16[] marks) {
+        public override long GetBlocksAffected(Level lvl, Vec3U16[] marks) {
             Vec3U16 p1 = marks[0], p2 = marks[1];
             return (p2.X - p1.X + 1) * (p2.Y - p1.Y + 1) * (p2.Z - p1.Z + 1);
         }
